+++ src/browseraction/BrowserAction.vue
@@ -6,6 +6,7 @@ import { BrowserStorageKey } from '@/browseraction/components/store/browser-stor
 import browser from 'webextension-polyfill'
 import { store } from '@/browseraction/components/store/store'
 import { ref } from 'vue'
+import { NO_TAB_GROUP_ID, loadTabGroups } from './components/logic/tabgroups'
 
 const isStoredValuesLoaded = ref(false)
 Promise.all([
@@ -14,14 +15,22 @@ Promise.all([
   browser.storage.local.get(BrowserStorageKey.random),
   browser.storage.local.get(BrowserStorageKey.reverse),
   browser.storage.local.get(BrowserStorageKey.preserve),
-  browser.storage.local.get(BrowserStorageKey.deduplicate)
+  browser.storage.local.get(BrowserStorageKey.deduplicate),
+  loadTabGroups(),
+  browser.storage.local.get(BrowserStorageKey.selectedTabGroupId)
 ]).then((data) => {
-  store.urlList = data[0][BrowserStorageKey.urlList] ?? ''
-  store.lazyLoadingChecked = data[1][BrowserStorageKey.lazyload] ?? false
-  store.loadInRandomOrderChecked = data[2][BrowserStorageKey.random] ?? false
-  store.loadInReverseOrderChecked = data[3][BrowserStorageKey.reverse] ?? false
-  store.preserveInputChecked = data[4][BrowserStorageKey.preserve] ?? false
-  store.deduplicateURLsChecked = data[5][BrowserStorageKey.deduplicate] ?? false
+  store.urlList = String(data[0][BrowserStorageKey.urlList] ?? '')
+  store.lazyLoadingChecked = Boolean(data[1][BrowserStorageKey.lazyload]) ?? false
+  store.loadInRandomOrderChecked = Boolean(data[2][BrowserStorageKey.random]) ?? false
+  store.loadInReverseOrderChecked = Boolean(data[3][BrowserStorageKey.reverse]) ?? false
+  store.preserveInputChecked = Boolean(data[4][BrowserStorageKey.preserve]) ?? false
+  store.deduplicateURLsChecked = Boolean(data[5][BrowserStorageKey.deduplicate]) ?? false
+  store.hasTabGroupSupport = Boolean(browser.tabGroups) ?? false
+  store.tabGroups = data[6]
+  store.selectedTabGroupId =
+    store.tabGroups.find(
+      (group) => group.id === Number(data[7][BrowserStorageKey.selectedTabGroupId])
+    )?.id ?? NO_TAB_GROUP_ID
 
   isStoredValuesLoaded.value = true
 })